Move new tracking classes into appropriate place
