+++ holypipette/gui/__init__.py
@@ -4,4 +4,5 @@ from .camera import *
 from .manipulator import *
 from .patch import *
 from .paramecium_device import *
+from .tracker import *
