+++ holypipette/interface/__init__.py
@@ -9,3 +9,4 @@ from .base import *
 from .camera import *
 from .pipettes import *
 from .patch import *
+from .tracker import *
